feat: 팔로우, 좋아요 validation 추가 (#46)
* feat: 카카오 OAuth 로그인 API 구현

* feat: SMS로 4자리 인증번호 API 구현

* fix: 채용공고 params 수정

* add: 팔로우, 좋아요 validation 추가

+++ src/main/java/com/example/demo/src/employment/EmploymentProvider.java
@@ -25,7 +25,7 @@ public class EmploymentProvider {
     }
 
     //채용공고 조회
-    public GetEmploymentRes getEmployments(Long userIdx, String country, String sort, int years1, int years2) throws BaseException {
+    public GetEmploymentRes getEmployments(Long userIdx, String country, String sort, Long years1, Long years2) throws BaseException {
         try {
             GetEmploymentRes getEmploymentRes = employmentDao.getEmployments(userIdx, country, sort, years1, years2);
             return getEmploymentRes;
